Deployment/requirements.txt
ultralytics>=8.0.0
flask
opencv-python
Pillow
